adds: * --no-cursors * --no-bell * --no-notifications
These options apply to both server and client. The client cannot override the server for these options. We now correctly grey out the options which are not available in the UI (and add the new ones to the UI).

Also, we now always capture the latest mouse cursor data, and only send the data if the client has the feature enabled: fixes a tiny bug where the client connects and the window is already meant to be showing a cursor (sending the latest in process_hello would not work in this case)

git-svn-id: https://xpra.org/svn/Xpra/trunk@1271 3bb7dfac-3a0b-4e04-842a-767bc560f471