XPRA_ANNOUNCE_TIMEOUT in ui_thread_watcher.py
The UI_thread_watcher identifies when the client is not responding, and
signals to the server.  By default, it also announces this in its log
output.  On some systems using OpenGL, this block can occur during the
normal enabling of monitor DPMS or a VT switch.  This leads to a large
number of these announcements

Provide a mechanism to reduce logging of ui_thread_watcher's detection
of by providing a configurable timeout, independent of the
ui_thread_watcher's detection threshold, XPRA_ANNOUNCE_TIMEOUT.

If set to 0, no such announcements will be logged.

Closes #3035

Signed-off-by: Antonio Russo <aerusso@aerusso.net>