Start server before the client
... if user invokes 'xpra start' or 'start-desktop' with
'attach' option.

For example, the command:

  xpra start --start=xeyes --attach=yes

gets hung because in daemon mode, either client or server
is started by 'do_run_mode'.

The following command works though:

  xpra start --start=xeyes --attach=yes --daemon=no

Now it works regardless of daemon mode.

Also checking if some server is running with the same arguments
is useless, because the end user explicitly called 'xpra start'
with clear intention to start the new session.

Signed-off-by: Vasyl Gello <vasek.gello@gmail.com>