Fix error that occurs when make_session_dir fallback already exists
If the usual session directory cannot be created, a temporary
directory is used as a fallback. However, there is no check to ensure
that the fallback does not already exist, leading to an error.

Xpra requires at least Python 3.6 now, so it is okay to use Python
3.3's `exist_ok` when calling `os.makedirs`.

Following this, we now always correct the owner of the directory for
consistency. This was previously skipped when the usual directory
(but not the fallback) already existed.