Don't close all open files when redirecting stdio
Conventional wisdom holds that a process about to background should first
close all its open file descriptors, so as to prevent holding on to any open
files it may have inherited from a careless parent.

This might have made sense back in the days of single-threaded programs with
no library dependencies out of its control, but it is most certainly not a
safe thing to do willy-nilly any more.

To wit, on my system, an nsswitch plugin starts a background thread that opens
some sockets on which it depends for its operation. Xpra then cheerfully,
mercilessly closes them all in preparation for daemonizing, and then makes
completely unrelated open calls that recycle those file descriptor numbers,
causing absolute havoc as the main program and the background thread share a
file descriptor they really did not expect to be sharing.

There is no possible way for the xpra server script to know this is happening
and gather a set of file descriptors to preserve. Instead we should close only
file descriptors we know we shouldn't have open any more. In xpra's case these
are only the stdio file descriptors that we are redirecting anyway.

Since we no longer violently close all file descriptors when daemonizing, we
no longer need to carefully track the extra files we open when creating a PAM
session, either; this change removes the code that does so.