#3626 correctly handle channel exec requests
this is not documented anywhere that I could find:
we have to first set the event, only then can we use send_exit_status() and sendall() / sendall_stderr()
otherwise the client ends up timing out when reading the stdout / stderr.
We also have to allow a new channel so that the client can run multiple commands.