Always try to find headers from /usr/include, /usr/local/include on !WIN32
setup.py uses sys.prefix to look for includes. This is where the
Python interpreter lives. However, that may be very different from
where the headers for system libraries live. For example, in a
virtualenv, sys.prefix points to the virtual environment, where the
presence of those headers is unlikely.

Make sure that /usr/include, /usr/local/include are always in the list
unless overridden by the INCLUDE_DIRS environment variable.

This is not a perfect heuristic by any means, but probably more likely
to find the libraries the user wants to find.