Client property updates should always include maximized
The maximized property is special - instead of
being stored on the server window object like
other properties, the server mirrors the value
back to the client every time it responds to a
state change message (see commit 26b99242)

To keep the maximized state set on a window, the
client has to send the maximized flag any time it
would otherwise send a property update. Without
this the server will default maximized to False
and the window will unmaximize itself after some
other event like changing the focus.