try harder to detect opengl properties
another change blacklisted 'llvmpipe' but when we're starting a server, we don't care about being able to run the opengl client code, we only want to query the opengl driver properties, so force it.
This fixes the 'OpenGL is not supported on this display' message shown in the server output.